+++ graphql/queries/image-generation/image-generation-task.ts
@@ -1,9 +1,12 @@
 import { gql } from "@/graphql/__generated__"
 
+/**
+ * 画像生成のタスク
+ */
 export const imageGenerationTaskQuery = gql(`
-query ImageGenerationTask($id: ID!) {
-  imageGenerationTask(id: $id) {
-    ...ImageGenerationTaskFields
-  }
+  query ImageGenerationTask($id: ID!) {
+    imageGenerationTask(id: $id) {
+      ...ImageGenerationTaskFields
+    }
 }
 `)
