+++ graphql/queries/viewer/viewer-folders.ts
@@ -1,5 +1,8 @@
 import { gql } from "@/graphql/__generated__"
 
+/**
+ * ログイン中のユーザのフォルダ
+ */
 export const viewerFoldersQuery = gql(`
   query ViewerFolders($offset: Int!, $limit: Int!) {
     viewer {
