+++ graphql/queries/viewer/viewer-liked-works.ts
@@ -1,5 +1,8 @@
 import { gql } from "@/graphql/__generated__"
 
+/**
+ * ログイン中のユーザのいいねした作品
+ */
 export const viewerLikedWorksQuery = gql(`
   query ViewerLikedWorks($offset: Int!, $limit: Int!) {
     viewer {
