+++ graphql/queries/viewer/viewer-message-threads.ts
@@ -1,5 +1,8 @@
 import { gql } from "@/graphql/__generated__"
 
+/**
+ * ログイン中のユーザのメッセージ
+ */
 export const viewerMessageThreadsQuery = gql(`
   query ViewerMessageThreads($offset: Int!, $limit: Int!) {
     viewer {
