+++ graphql/queries/viewer/viewer-muted-tags.ts
@@ -1,5 +1,8 @@
 import { gql } from "@/graphql/__generated__"
 
+/**
+ * ログイン中のユーザのミュートしたタグ
+ */
 export const viewerMutedTagsQuery = gql(`
   query ViewerMutedTags($offset: Int!, $limit: Int!) {
     viewer {
