+++ graphql/queries/viewer/viewer-muted-users.ts
@@ -1,5 +1,8 @@
 import { gql } from "@/graphql/__generated__"
 
+/**
+ * ログイン中のユーザのミュートのユーザ
+ */
 export const viewerMutedUsersQuery = gql(`
   query ViewerMutedUsers($offset: Int!, $limit: Int!) {
     viewer {
