+++ graphql/queries/viewer/viewer-payments.ts
@@ -1,5 +1,8 @@
 import { gql } from "@/graphql/__generated__"
 
+/**
+ * ログイン中のユーザの決済履歴
+ */
 export const viewerPaymentsQuery = gql(`
   query ViewerPayments {
     viewer {
