+++ graphql/queries/viewer/viewer-support-messages.ts
@@ -1,5 +1,8 @@
 import { gql } from "@/graphql/__generated__"
 
+/**
+ * ログイン中のユーザのシリーズのサポートメッセージ
+ */
 export const viewerSupportMessagesQuery = gql(`
   query ViewerSupportMessages($offset: Int!, $limit: Int!) {
     viewer {
