+++ graphql/queries/viewer/viewer-works.ts
@@ -1,5 +1,8 @@
 import { gql } from "@/graphql/__generated__"
 
+/**
+ * ログイン中のユーザの作品
+ */
 export const viewerWorksQuery = gql(`
   query ViewerWorks($offset: Int!, $limit: Int!) {
     viewer {
